Move atom feeds functionality to peachjam and refactor import references
